refactor: member Controller Command, Query분리
